(svn r8705) -Codechange: Increased the number of airport blocks to 64. This involves changing the enum of airport blocks to a static const uint64 as SOME platforms do not support 64-bit enums

+++ src/airport_movement.h
@@ -9,7 +9,7 @@
 typedef struct AirportFTAbuildup {
 	byte position; // the position that an airplane is at
 	byte heading;  // the current orders (eg. TAKEOFF, HANGAR, ENDLANDING, etc.)
-	uint32 block;  // the block this position is on on the airport (st->airport_flags)
+	uint64 block;  // the block this position is on on the airport (st->airport_flags)
 	byte next;     // next position from this position
 } AirportFTAbuildup;
